requirements.txt
Flask
scikit-learn
nltk
python-dotenv
gunicorn
